[add_test] add move semantic, refactoring

+++ simple_vector_lib/include/utils.hpp
@@ -19,7 +19,7 @@ namespace utils
     inline uint32_t calculate_capacity(const uint32_t& size)
     {
         // @todo: fix? diff formula or round
-        return static_cast<uint32_t>(size * 1.5f);
+        return static_cast<uint32_t>(static_cast<float>(size) * 1.5f);
     }
 };  // namespace utils
